Fix debug mode display
Currently, setting `APP_DEBUG=true` in my `.env` file results in the dashboard saying debug mode is "Disabled". Setting it to `false` makes the dashboard display "Enabled". 
Simply reversing the order here fixes it so it properly displays "Disabled" when APP_DEBUG is set to false, and "Enabled" when it is set to true.